42push

+++ ft_hexadecimal_uppercase.c
@@ -6,7 +6,7 @@
 /*   By: hirwatan <hirwatan@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/02 21:44:35 by hirwatan          #+#    #+#             */
-/*   Updated: 2024/11/06 20:17:26 by hirwatan         ###   ########.fr       */
+/*   Updated: 2024/11/07 19:16:58 by hirwatan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 #include <stdarg.h>
 #include <unistd.h>
 
-static int	ft_putnbr_base_u(unsigned long nbr, char *base)
+static int	ft_putnbr_base_u(unsigned int nbr, char *base)
 {
 	int		len;
 	char	c;
@@ -29,8 +29,8 @@ static int	ft_putnbr_base_u(unsigned long nbr, char *base)
 
 int	ft_hexadecimal_uppercase(unsigned int args)
 {
-	int len;
-	char c;
+	int		len;
+	char	c;
 
 	len = 0;
 	if (args == 0)
@@ -38,6 +38,6 @@ int	ft_hexadecimal_uppercase(unsigned int args)
 		c = '0';
 		return (write(1, &c, 1));
 	}
-	len += ft_putnbr_base_u((unsigned long)args, "0123456789ABCDEF");
+	len += ft_putnbr_base_u((unsigned int)args, "0123456789ABCDEF");
 	return (len);
 }
